+++ ios/CHANGELOG.md
@@ -24,6 +24,8 @@ Line wrap the file at 100 chars.                                              Th
 ## Unreleased
 ### Fixed
 - Broken DAITA settings view on iOS 15.
+### Changed
+- Move Changelog to settings and add an in-app notification banner for an update app
 
 ## [2025.1 - 2025-01-14]
 ### Added